refactor neon-cli
