+++ evm_loader/cli/src/commands/create_ether_account.rs
@@ -17,6 +17,7 @@ use crate::{
     Config,
     NeonCliResult,
     rpc::{Rpc, NodeClient},
+    account_storage::make_solana_program_address,
 };
 
 
@@ -25,7 +26,7 @@ pub fn execute (
     ether_address: &H160,
 ) -> NeonCliResult {
 
-    let (solana_address, nonce) = crate::make_solana_program_address(ether_address, &config.evm_loader);
+    let (solana_address, nonce) = make_solana_program_address(ether_address, &config.evm_loader);
     debug!("Create ethereum account {} <- {} {}", solana_address, hex::encode(ether_address), nonce);
 
     let create_account_v03_instruction = Instruction::new_with_bincode(
