+++ evm_loader/cli/src/commands/create_program_address.rs
@@ -2,13 +2,14 @@ use evm::{H160};
 
 use crate::{
     Config,
+    account_storage::make_solana_program_address,
 };
 
 pub fn execute (
     config: &Config,
     ether_address: &H160,
 ) {
-    let (solana_address, nonce) = crate::make_solana_program_address(ether_address, &config.evm_loader);
+    let (solana_address, nonce) = make_solana_program_address(ether_address, &config.evm_loader);
     println!("{} {}", solana_address, nonce);
 }
