+++ evm_loader/cli/src/commands/deposit.rs
@@ -18,6 +18,7 @@ use crate::{
     Config,
     NeonCliResult,
     rpc::{Rpc, NodeClient},
+    account_storage::make_solana_program_address,
 };
 
 /// Executes subcommand `deposit`.
@@ -26,7 +27,7 @@ pub fn execute(
     amount: u64,
     ether_address: &H160,
 ) -> NeonCliResult {
-    let (ether_pubkey, nonce) = crate::make_solana_program_address(ether_address, &config.evm_loader);
+    let (ether_pubkey, nonce) = make_solana_program_address(ether_address, &config.evm_loader);
 
     let mut instructions = Vec::with_capacity(2);
